Convert time-summary to typescript
Fix types for time-summary
